requirements.txt
Flask
opencv-python-headless
requests
numpy
python-dotenv